fix: LoginId 반환 메서드 Controller로 이동
